[Fix/#1] 나연엄마의 코리 반영!!!
- extension View와 Text 파일로 합치기
- font함수 네이밍 변경
