+++ Gongbaek_iOS/Gongbaek_iOS/Global/Extension/Text+.swift
@@ -8,7 +8,16 @@
 import SwiftUI
 
 extension Text {
-    func font(_ font: FontFamily) -> some View {
+    func pretendardFont(_ font: FontFamily) -> some View {
         self.applyFont(font)
+    }
+    
+    func applyFont(_ font: FontFamily) -> some View {
+        let lineSpacingValue = font.lineHeight - font.fontSize
+        
+        return self.font(.pretendard(font))
+            .tracking(font.letterSpacing)
+            .lineSpacing(lineSpacingValue)
+            .padding(.vertical, lineSpacingValue/2)
     }
 }
