--- Gongbaek_iOS/Gongbaek_iOS/Global/Extension/View+.swift
@@ -1,19 +1,0 @@
-//
-//  View+.swift
-//  Gongbaek_iOS
-//
-//  Created by 김희은 on 1/9/25.
-//
-
-import SwiftUI
-
-extension View {
-    func applyFont(_ font: FontFamily) -> some View {
-        let lineSpacingValue = font.lineHeight - font.fontSize
-        
-        return self.font(.pretendard(font))
-            .tracking(font.letterSpacing)
-            .lineSpacing(lineSpacingValue)
-            .padding(.vertical, lineSpacingValue/2)
-    }
-}
